refactor: Rename `…/event-detection/event-types/customAuthorizer.js`
