+++ sdk-js/src/lib/event-detection/event-types/custom-authorizer.js
