+++ sdk-js/src/lib/event-detection/index.js
@@ -3,7 +3,7 @@
 const alexaSkill = require('./event-types/alexa-skill');
 const apiGateway = require('./event-types/api-gateway');
 const apiGatewayV2 = require('./event-types/api-gateway-v2');
-const customAuthorizer = require('./event-types/customAuthorizer');
+const customAuthorizer = require('./event-types/custom-authorizer');
 const cloudFront = require('./event-types/cloud-front');
 const cloudwatchEvent = require('./event-types/cloudwatch-event');
 const cloudwatchLogs = require('./event-types/cloudwatch-log');
